shell/: Redo the EGL window creation
This defers creating the EGL window/surface right after the
xdg_surface::configure ack to be able to create the window with the
dimensions from the beginning, rather than attempting to resize
aftwards, which risks of getting a protocol violations because we're
still attaching+commiting a buffer with a previous although we did
ack'ed it.

This basically moves the EGL window creation right after that initial
configure and allows starting up as maximized/fullscreen.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>