shell: Do not issue always the ready request
Rather than sending the ready request always, just do it when we've
found at least a background type of surface.

This way we can have a flutter instance as a regular application but
also as a shell client.

This is a work-around for the time being, a much more suitable approach
would be to have a mechanism that tells us when the client has finished
loading or has it's first frame swapped (or something similar to that).

Note that we need at least a background surface so not finding one
assumes that we're a regular application (WINDOW_NORMAL).

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>