key_event: ignore some keys about cursol moving to avoid weird behavior
When the keys about cursol moving are pushed in TextField,
since the keys are handled twice i.e. in both of key_event and text_input plugins,
the behavior of a cursol will be weird.

Maybe, the correct flow is the following.

- When a key is handled as Logical key events in apps, such a key should
  be ignored by text_input.
- When a key is not handled in apps, such a key should be handled by
  text_input.

However, I'm not sure how to know if a key is handled in apps.
So, as a workaround, I let keys about cursol moving ignored in
key_event.

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>