Staged release (#58)
* Region leak fix (#55)

-destroy region after use since there are copy semantics involved.

Signed-off-by: Joel Winarske <joel.winarske@woven-planet.global>

* Pixel Ratio fix (#56)

-The Flutter engine handles scaling of pointer events
 no need to scale externally.  Remove prior handling.
-add comp_region test case to source tree

Signed-off-by: Joel Winarske <joel.winarske@woven-planet.global>

* Fix KeyEvent and TextInput to work exclusively (#54)

If the both of KeyEvent and TextInput work at same time for a same
widget(same focus node), the behavior of an input cursor of TextInput
will become weird i.e. a input cursor moves at a head of a line per
every keyinputs.

This issue can be triggered by a flutter code like the following.

```
Container(
    padding: const EdgeInsets.all(16.0),
    child: Focus(
        autofocus: true,
        onKey: (FocusNode node, RawKeyEvent event) {
            var keyLabel = event.logicalKey.keyLabel.isNotEmpty
                ? event.logicalKey.keyLabel
                : event.physicalKey.debugName.toString();

            var retString = "unknown";
            if (event is RawKeyDownEvent) {
                if (event.repeat) {
                    retString = "KeyRepeat_" + keyLabel;
                } else {
                    retString = "KeyDown_" + keyLabel;
                }
            } else if (event is RawKeyUpEvent) {
                retString = "KeyUp_" + keyLabel;
            }

            if (retString == "") {
                return KeyEventResult.ignored;
            } else {
                setState(() {
                    _text += '[' + retString + ']';
                });
                return KeyEventResult.handled;
            }
        },
        child: Column(
            children: <Widget>[
                TextField(
                    controller: TextEditingController(text: _current),
                    maxLines: 20,
                    style: const TextStyle(color: Colors.black),
                    onChanged: (text) {
                        _current = text;
                    },
                ),
                Text(_text),
            ],
       )
    ),
),
```

If the above widget runs with ivi-homescreen, any keyevent will be
handled by the both of `Text` widget and `TextField` widget.

However, in other platform e.g. linux desktop, windows desktop,
when `onKey()` returns `KeyEventResult.handled` i.e. a key input is
handled as RawKeyEvent, a key input will not be handled by `TextField`.

This means just one of keyevent handling and keyinput handling needs to work
in the embedder internal.
i.e. TextInput and KeyEvent need to work exclusively for each.

To let KeyEvent and TextInput work exclusively, the sequence like the
following is required.

- Wayland send a key event to ivi-homescreen
- ivi-homescreen receives a key event
- KeyEvent send a key event to a flutter app.
- A flutter app send the reply.
- KeyEvent receives the reply and parses it.
- When a key event is ignored, KeyEvent invokes the key handler of
  TextInput.
- When a key event ia handled, KeyEvent doesn't invoke the key handler
  of TextInput.

Currently, KeyEvent cannot receive any reply from a flutter app.
So, I added the support of BinaryReply for BasicMessageChannel.Send().
Due to this, ivi-homescreen can get a result whether if a key event has
been handled on the side of a flutter app.

And also, currently, the key handler of TextInput and KeyEvent
(`keyboard_handle_key()`) are invoked sequencially.
So, I changed the design of TextInput to provide KeyEvent the callback to handle
a key event for KeyEvent and not to invoke the handler from other parts.
(This callback is called "DelegateHandleKey" internal.)

However, if only TextInput is enabled, the handler will be invoked
directly from `keyboard_handle_key()` as same as the old implementation.

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>

* Fixes typo in touch_handle_motion (#57)

Signed-off-by: Joel Winarske <joel.winarske@woven-planet.global>

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>
Co-authored-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>