shell: main.cc: handle std::stol exception
Currently, the exception of std::stol is not handled, so, when
specifying too large number as an argument, the following exception will
be raised.

```
terminating with uncaught exception of type std::out_of_range: stol: out of range
Aborted (core dumped)
```

So, I added the code to handle the exception of std::stol and then,
I replaced IsNumber()'s check with std::invalid_argument exception
handling.

And also, due to it, I changed the code to allow the hex-style like `0x1f`
and the octet style like `031` as an argument of --a option.

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>