Update damage area
Rather than re-using the initial values from the config, update the
values we are getting from the compositor to match the windows'
dimensions.

This fixes some artifacts we're seeing when re-drawing with swap buffer
with damage and most likely would have caused issues on other
compositors when resizing the window to a higher value that the one
provided in the configuration file.

Bug-AGL: SPEC-4938
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>