wayland_egl: fix build error of -fpermissive
This part causes the following build error.
So, I fixed it with the explicit conversion with using const_cast.
The type declaration of `const auto` is due to clang-tidy+reformat(i.e.
statics analytics), so, I kept it as is.

```
/home/akiyama/work/flutter/ivi-homescreen/shell/backend/wayland_egl.cc:
In lambda function:
/home/akiyama/work/flutter/ivi-homescreen/shell/backend/wayland_egl.cc:116:74:
error: invalid conversion from ‘std::array<int, 4>::const_pointer’ {aka
‘const int*’} to ‘int*’ [-fpermissive]
  116 |                       b->GetDisplay(), b->m_egl_surface,
      frame_rects.data(), 1);
            |
            ~~~~~~~~~~~~~~~~^~
                  |
                  |
                        |
                        std::array<int, 4>::const_pointer {aka const
                        int*}
```

Signed-off-by: Kazuyoshi Akiyama <kazuyoshi.akiyama@woven-planet.global>