Fix AGL shell activation area behavior
Commit 0e52d4d1 accidentally broke using an activation area
configuration with either x or y set to 0.  Rework the logic
to use the specified configuration area if any of the position
or size values are non-zero.  This restores expected behavior
in the downstream AGL demos.  As well, drop the fallback
default activation area configuration, as it seems better to
not configure it at all if it has not been explicitly
configured.

Signed-off-by: Scott Murray <scott.murray@konsulko.com>